GBPTreeRecoveryTest uses ephemeral fs rule and test-directory rules
